+++ CHANGELOG.md
@@ -9,7 +9,8 @@ CHANGELOG
 
 ## Upcoming version ##
 
-* 🛠 No longer enforce location permission ([#](https://github.com/tankste/app/issues/))
+* ➕ Add date of last price change to details ([#2](https://github.com/tankste/app/issues/2))
+* 🛠 No longer enforce location permission ([#6](https://github.com/tankste/app/issues/6))
 * ➕ Add settings option for changing the map provider ([#51](https://github.com/tankste/app/issues/51))
 * ➕ Add sponsor page to settings ([#56](https://github.com/tankste/app/issues/49))
 * 🛠 Use tankste! backend and remove Tankerkoenig usage ([#49](https://github.com/tankste/app/issues/49))